Change styling of tags in input field when comma is entered
